requirements.txt
asgiref==3.5.2
Django~=3.2
django-ninja==0.19.0
pydantic==1.9.1
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.3.0
